refactor: Move relative code inside `AddModal` folder
no code change. The AddModal folder will be extracted in cozy-ui
after that
