+++ src/components/AddModal/ContactForm/FieldInputWrapper.jsx
@@ -2,10 +2,9 @@ import React from 'react'
 
 import TextField from 'cozy-ui/transpiled/react/TextField'
 
-import { FieldInputWrapperPropTypes } from './fieldInputType'
-
-import TextFieldCustomLabelSelect from '@/components/Form/TextFieldCustomLabelSelect'
-import TextFieldSelect from '@/components/Form/TextFieldSelect'
+import TextFieldCustomLabelSelect from './TextFieldCustomLabelSelect'
+import TextFieldSelect from './TextFieldSelect'
+import { FieldInputWrapperPropTypes } from '../types'
 
 // component used to flatten props to ensure compatibility
 // between Field from react-final-form and TextField from Mui
