+++ src/components/ContentResult.spec.jsx
@@ -1,8 +1,9 @@
 import { render, screen, fireEvent, act, waitFor } from '@testing-library/react'
 import React from 'react'
 
+import { johnDoeContact } from './AddModal/mocks'
 import ContentResult from './ContentResult'
-import { groups, contactWithGroup, johnDoeContact } from '../helpers/testData'
+import { groups, contactWithGroup } from '../helpers/testData'
 import enLocale from '../locales/en.json'
 import AppLike from '../tests/Applike'
