+++ src/helpers/contacts.spec.js
@@ -3,14 +3,14 @@ import MockDate from 'mockdate'
 import {
   getConnectedAccounts,
   normalizeFields,
-  updateIndexFullNameAndDisplayName,
   harmonizeAndSortByFamilyNameGivenNameEmailCozyUrl,
   reworkContacts,
-  makeFormattedAddressWithSubFields,
-  getFormattedAddress,
   makeContactWithIdentitiesAddresses
 } from './contacts'
-import { johnDoeContact, johnDoeContactWithRelated } from './testData'
+import {
+  johnDoeContact,
+  johnDoeContactWithRelated
+} from '../components/AddModal/mocks'
 
 const MOCKED_DATE = '2018-01-01T12:00:00.210Z'
 
@@ -147,21 +147,6 @@ describe('normalizeFields', () => {
     ]
 
     expect(normalizeFields(johnDoeContactWithRelated)).toEqual(expected)
-  })
-})
-
-describe('updateIndexFullNameAndDisplayName', () => {
-  it('should returns a contact with new attributes', () => {
-    const expected = {
-      ...johnDoeContact,
-      displayName: 'John J. Doe',
-      fullname: 'John J. Doe',
-      indexes: {
-        byFamilyNameGivenNameEmailCozyUrl:
-          'doejohnjohn.doe@cozycloud.ccjohndoe.mycozy.cloud'
-      }
-    }
-    expect(updateIndexFullNameAndDisplayName(johnDoeContact)).toEqual(expected)
   })
 })
 
@@ -369,146 +354,6 @@ describe('reworkContacts', () => {
         contactsWithNoIndexes
       )
     ).toEqual(expected)
-  })
-})
-
-describe('getFormattedAddress', () => {
-  describe('with connector address format', () => {
-    it('should return full formatted address', () => {
-      const addressMock = {
-        city: 'Cambridge',
-        country: 'Russian Federation',
-        postcode: '16862',
-        street: '38 Taylor Street'
-      }
-
-      const res = getFormattedAddress(
-        addressMock,
-        jest.fn(() => '38 Taylor Street, 16862 Cambridge, Russian Federation')
-      )
-
-      expect(res).toBe('38 Taylor Street, 16862 Cambridge, Russian Federation')
-    })
-    it('should return formatted address if only "postcode" & "city" values are defined', () => {
-      const addressMock = {
-        city: 'Cambridge',
-        country: undefined,
-        postcode: '16862',
-        street: undefined
-      }
-      const res = getFormattedAddress(
-        addressMock,
-        jest.fn(() => ' , 16862 Cambridge, ')
-      )
-
-      expect(res).toBe('16862 Cambridge')
-    })
-    it('should return formatted address if "postcode" & "city" values are undefined', () => {
-      const addressMock = {
-        city: 'Cambridge',
-        country: undefined,
-        postcode: undefined,
-        street: '38 Taylor Street'
-      }
-      const res = getFormattedAddress(
-        addressMock,
-        jest.fn(() => '38 Taylor Street,  , Cambridge')
-      )
-
-      expect(res).toBe('38 Taylor Street, Cambridge')
-    })
-    it('should return formatted address if "postcode" & "city" values are undefined', () => {
-      const addressMock = {
-        city: undefined,
-        country: undefined,
-        postcode: undefined,
-        street: undefined
-      }
-      const res = getFormattedAddress(
-        addressMock,
-        jest.fn(() => ' ,  , ')
-      )
-
-      expect(res).toBe('')
-    })
-  })
-
-  describe('with manual address format', () => {
-    it('should return full formatted address', () => {
-      const addressMock = {
-        formattedAddress: '38 Taylor Street 16862 Cambridge Russian Federation'
-      }
-
-      const res = getFormattedAddress(
-        addressMock,
-        jest.fn(() => '')
-      )
-
-      expect(res).toBe('38 Taylor Street 16862 Cambridge Russian Federation')
-    })
-  })
-})
-
-describe('makeFormattedAddressWithSubFields', () => {
-  it('should return full formatted address', () => {
-    const subFieldsStateMock = [
-      { name: 'address[0].addressnumber', value: '10' },
-      { name: 'address[0].addressstreet', value: 'rue du test' },
-      { name: 'address[0].addresscode', value: '75056' },
-      { name: 'address[0].addresscity', value: 'Paris' },
-      { name: 'address[0].addresscountry', value: 'France' }
-    ]
-    const res = makeFormattedAddressWithSubFields(
-      subFieldsStateMock,
-      jest.fn(() => '10 rue du test, 75056 Paris, France')
-    )
-
-    expect(res).toBe('10 rue du test, 75056 Paris, France')
-  })
-  it('should return formatted address if only "code" & "city" values are defined', () => {
-    const subFieldsStateMock = [
-      { name: 'address[0].addressnumber', value: undefined },
-      { name: 'address[0].addressstreet', value: undefined },
-      { name: 'address[0].addresscode', value: '75056' },
-      { name: 'address[0].addresscity', value: 'Paris' },
-      { name: 'address[0].addresscountry', value: undefined }
-    ]
-    const res = makeFormattedAddressWithSubFields(
-      subFieldsStateMock,
-      jest.fn(() => ' , 75056 Paris, ')
-    )
-
-    expect(res).toBe('75056 Paris')
-  })
-  it('should return formatted address if "code" & "city" values are undefined', () => {
-    const subFieldsStateMock = [
-      { name: 'address[0].addressnumber', value: '10' },
-      { name: 'address[0].addressstreet', value: 'rue du test' },
-      { name: 'address[0].addresscode', value: undefined },
-      { name: 'address[0].addresscity', value: undefined },
-      { name: 'address[0].addresscountry', value: 'France' }
-    ]
-    const res = makeFormattedAddressWithSubFields(
-      subFieldsStateMock,
-      jest.fn(() => '10 rue du test,  , France')
-    )
-
-    expect(res).toBe('10 rue du test, France')
-  })
-  it('should return formatted address if all values are undefined', () => {
-    const subFieldsStateMock = [
-      { name: 'address[0].addressnumber', value: undefined },
-      { name: 'address[0].addressstreet', value: undefined },
-      { name: 'address[0].addresscode', value: undefined },
-      { name: 'address[0].addresscity', value: undefined },
-      { name: 'address[0].addresscountry', value: undefined }
-    ]
-    const res = makeFormattedAddressWithSubFields(
-      subFieldsStateMock,
-      jest.fn(() => ' ,  , ')
-    )
-
-    expect(res).toBe('')
   })
 })
