+++ src/styles/index.styl
@@ -1,3 +1,2 @@
 @require './contacts.styl'
-@require './contactForm.styl'
 @require './contactImportation.styl'
